Metal VertexDescriptor.

+++ Texture2d/iOS-Metal/Sprite/Core/Sprite.h
@@ -13,7 +13,7 @@
 
 @interface Sprite : NSObject
 
-@property(atomic,assign) CGFloat angle, scale, transX, transY;
+@property(atomic,assign) float angle, scale, transX, transY;
 @property(atomic,strong) NSString* textureImagePath;
 
 -(instancetype)initWithDevice:(id<MTLDevice>)device;
